General cleanup of learner.cpp.
